Extract server creation for reuse
